requirements.txt
llama-index
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-vector-stores-postgres
llama-index-storage-docstore-postgres
llama-index-storage-docstore-postgres
python-dotenv
psycopg2-binary
crewai
crewai-tools
ragas
datasets
rank_bm25
llama-index-retrievers-bm25
sentencepiece
docling-ibm-models[mlx]
fastapi
uvicorn
openinference-instrumentation-crewai
arize-phoenix-otel
openinference-instrumentation-crewai
opentelemetry-sdk
openinference-instrumentation-litellm